LPS-184766 HttpServiceRuntimeController: Move HttpService related logic to HttpServiceImpl, prepare for next
