+++ modules/apps/static/portal-osgi-web/portal-osgi-web-http-servlet-impl/src/main/java/com/liferay/portal/osgi/web/http/servlet/internal/activator/HttpServletImplBundleActivator.java
@@ -223,7 +223,7 @@ public class HttpServletImplBundleActivator implements BundleActivator {
 			HttpServiceRuntimeController httpServiceRuntimeController =
 				new HttpServiceRuntimeController(
 					_bundleContext, _bundleContext, servletContext,
-					targetFilter, Collections.unmodifiableMap(attributesMap));
+					Collections.unmodifiableMap(attributesMap));
 
 			proxyServlet.setHttpServiceRuntimeController(
 				httpServiceRuntimeController);
@@ -241,7 +241,8 @@ public class HttpServletImplBundleActivator implements BundleActivator {
 					).build()),
 				_bundleContext.registerService(
 					HttpService.class,
-					new HttpServiceFactory(httpServiceRuntimeController),
+					new HttpServiceFactory(
+						httpServiceRuntimeController, targetFilter),
 					HashMapDictionaryBuilder.putAll(
 						attributesMap
 					).build()),
